+++ content-quote.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * The template for displaying posts in the Quote post format on index and archive pages.
- * @version 6.4.1
+ * @version 6.4.1.1
  */
 ?>
 	<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>