+++ footer.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * The template for displaying the footer.
- * @version 6.4.1
+ * @version 6.4.1.1
  */
 ?>
             </div><!-- .grid-x .grid-margin-x -->