+++ js/theme-functions.js
@@ -1,4 +1,4 @@
-/* @version 6.4.1 */
+/* @version 6.4.1.1 */
 jQuery(document).foundation();
 
 	// contain-to-grid functionality