+++ single.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * The Template for displaying all single posts.
- * @version 6.4.1
+ * @version 6.4.1.1
  */
 get_header(); ?>
 	<div id="content" class="small-12 large-8 cell" role="main">